Creacion de la clase Auto
