--- Garcia-Ruiz-Villa-Agencia3/Cliente.java
@@ -1,11 +1,0 @@
-/**
- *Clase cliente que modela a un cliente interesado en comprar en nuestra agencia
- *@author: Lechuzas 2.0
- */
-
-
-public class Cliente{
-    //código
-
-
-}
