--- Pereyra-Bellon-Ferrary/Auto.java
@@ -1,4 +1,0 @@
-public class Auto{
-
-
-}
